FIX: para que la pausa se avise en la pantalla combate
