+++ bombava-frontend/src/componentes/botones/MenuPausa.jsx
@@ -3,7 +3,6 @@ import { useNavigate } from 'react-router-dom';
 import { peticionAbandonarPartida, peticionPausarPartida } from '../../utils/socket';
 import { cargarEstadoPartida, eliminarEstadoPartida } from '../../services/gameApi';
 import '../../styles/MenuPausa.css';
-import { socket } from '../../utils/socket';
 import { notification } from '../../services/notificationService';
 
 function MenuPausa() {
@@ -12,43 +11,6 @@ function MenuPausa() {
 
     const navigate = useNavigate();
 
-    // Enseña el cuadro para decidir si aceptar la solictud de pausa o no
-    const [cuadroDecidirSiPausarVisible, setCuadroDecidirSiPausarVisible] = useState(false);
-
-
-    // Configurar los listeners del socket al llegar a la pantalla
-    useEffect(() => {
-
-        //Recibe la señal de que han pedido pausar con el nombre del usuario que la ha aceptado
-        const handlePauseRequested = (payload) => {
-            notification.info(payload.from + "ha pedido pausar.");
-            setCuadroDecidirSiPausarVisible(true);
-        };
-
-        // Se ha pausado la partida, o yo he aceptado o el otro ha aceptado pausar
-        const handleMatchPaused = (payload) => {
-            notification.info(payload.message);
-            navigate('/menuInicial');
-        };
-
-        // Si rechaza el otro pausar la partida
-        const handlePauseReject = (payload) => {
-            notification.info(payload.message);
-            cerrarMenu();
-        };
-
-        // Escuchamos los siguientes eventos:
-        socket.on('match:pause_requested', handlePauseRequested);
-        socket.on('match:paused', handleMatchPaused);
-        socket.on('match:pause_rejected', handlePauseReject);
-
-        // Cerramos los listeners al salir de la pantalla
-        return () => {
-            socket.off('match:pause_requested', handlePauseRequested);
-            socket.off('match:paused', handleMatchPaused);
-            socket.off('match:pause_rejected', handlePauseReject);
-        };
-    }, []);
 
     // Función para abrir el menú
     const abrirMenu = () => {
@@ -80,28 +42,11 @@ function MenuPausa() {
         if (estado) {
             const matchId = estado.matchInfo.matchId;
             peticionPausarPartida(matchId);
+            notification.info("Peticion de pausa enviada");
+            cerrarMenu();
         }
         //Una vez ya he pedido al backend que le diga si quiere pausar la partida al otro jugador espero a que llegue
         //confirmación de ese jugador en handlePauseAccept para pausar o no hace nada
-    };
-
-    const aceptarPausa = () =>{
-        const estado = cargarEstadoPartida();
-        if (estado) {
-            const matchId = estado.matchInfo.matchId;
-            socket.emit('match:pause_accept',matchId);
-            setCuadroDecidirSiPausarVisible(false);
-        }
-        
-    };
-
-    const rechazarPausa = () =>{
-        const estado = cargarEstadoPartida();
-        if (estado) {
-            const matchId = estado.matchInfo.matchId;
-            socket.emit('match:pause_reject',matchId);
-            setCuadroDecidirSiPausarVisible(false);
-        }
     };
 
     return (
@@ -136,23 +81,6 @@ function MenuPausa() {
                             </button>
                         </div>
                     </div>
-                    {/* CUADRO PARA EL JUGADOR QUE RECIBE LA PETICIÓN */}
-                    {cuadroDecidirSiPausarVisible && (
-                        <div className="menu-pausa-fondo">
-                            <div className="menu-pausa">
-                                <h2>El oponente quiere pausar la partida</h2>
-                                <p>¿Aceptas guardar el estado actual y salir?</p>
-                                <div className="menu-botones">
-                                    <button className="btn-menu btn-continuar" onClick={aceptarPausa}>
-                                        Aceptar y Salir
-                                    </button>
-                                    <button className="btn-menu btn-abandonar" onClick={rechazarPausa}>
-                                        Rechazar
-                                    </button>
-                                </div>
-                            </div>
-                        </div>
-                    )}
                 </div>
                 
             )}
